Move common schema a layer up
