+++ src/web/rest/schema/images.ts
@@ -1,4 +1,4 @@
-import {joi} from "./common";
+import {joi} from "../../schema/common";
 
 export const CreateImageRequestSchema = joi.object({
     title: joi.string().required(),
